+++ .gitignore
@@ -59,3 +59,7 @@ typings/
 
 # next.js build output
 .next
+
+tempCodeRunnerFile\.js
+
+testing\.js
